Drop dry-v compatibility < 1.x
