+++ .travis.yml
@@ -3,6 +3,7 @@ cache: bundler
 bundler_args: --without benchmarks tools
 rvm:
   - ruby-head
+  - 2.6
   - 2.5
   - 2.4
 env:
@@ -10,17 +11,10 @@ env:
   - "DRY_VALIDATION='~> 1.2.0'"
   - "DRY_VALIDATION='~> 1.1.0'"
   - "DRY_VALIDATION='~> 1.0.0'"
-  - "DRY_VALIDATION='~> 0.13.0'"
-  - "DRY_VALIDATION='~> 0.12.0'"
 matrix:
   fast_finish: true
   allow_failures:
     - rvm: ruby-head
-  include:
-    - { rvm: 2.2, env: "DRY_VALIDATION='~> 0.11.0'" }
-    - { rvm: 2.3, env: "DRY_VALIDATION='~> 0.12.0'" }
-    - { rvm: 2.3, env: "DRY_VALIDATION='~> 0.13.0'" }
-
 
 # maybe move this to slack in a private channel
 # notifications:
