+++ test/composition_test.rb
