+++ test/contract_test.rb
