+++ test/inherit_test.rb
