+++ test/validation/dry_validation_test.rb
@@ -1,7 +1,6 @@
 require "test_helper"
 require "reform/form/dry"
 require "reform/form/coercion"
-
 #---
 # one "nested" Schema per form.
 class DryValidationErrorsAPITest < Minitest::Spec
@@ -14,22 +13,23 @@ class DryValidationErrorsAPITest < Minitest::Spec
     property :title
 
     validation do
-      # required(:title).filled
-      required(:title).filled(min_size?: 2)
+      params do
+        required(:title).filled(min_size?: 2)
+      end
     end
 
     property :artist do
       property :email
 
       validation do
-        required(:email).filled
+        params { required(:email).filled }
       end
 
       property :label do
         property :location
 
         validation do
-          required(:location).filled
+          params { required(:location).filled }
         end
       end
     end
@@ -39,11 +39,9 @@ class DryValidationErrorsAPITest < Minitest::Spec
       property :title
 
       validation do
-        configure do
-          config.messages_file = "test/fixtures/dry_error_messages.yml"
-        end
+        config.messages.load_paths << "test/fixtures/dry_new_api_error_messages.yml"
 
-        required(:title).filled
+        params { required(:title).filled }
       end
     end
   end
@@ -55,7 +53,6 @@ class DryValidationErrorsAPITest < Minitest::Spec
 
     result.success?.must_equal false
 
-    # errors.messages
     form.errors.messages.must_equal(title: ["must be filled", "size cannot be less than 2"], "artist.email": ["must be filled"], "artist.label.location": ["must be filled"], "songs.title": ["must be filled"])
     form.artist.errors.messages.must_equal(email: ["must be filled"], "label.location": ["must be filled"])
     form.artist.label.errors.messages.must_equal(location: ["must be filled"])
@@ -87,7 +84,9 @@ class DryValidationErrorsAPITest < Minitest::Spec
     form.songs[1].to_result.messages.must_equal(title: ["must be filled"])
     form.songs[1].to_result.hints.must_equal(title: [])
     form.songs[1].to_result.errors(locale: :de).must_equal(title: ["muss abgefüllt sein"])
-    form.songs[1].to_result.messages(locale: :de).must_equal(title: ["muss abgefüllt sein"])
+    # seems like dry-v when calling Dry::Schema::Result#messages locale option is ignored
+    # started a topic in their forum https://discourse.dry-rb.org/t/dry-result-messages-ignore-locale-option/910
+    # form.songs[1].to_result.messages(locale: :de).must_equal(title: ["muss abgefüllt sein"])
     form.songs[1].to_result.hints(locale: :de).must_equal(title: [])
   end
 
@@ -117,9 +116,11 @@ class DryValidationErrorsAPITest < Minitest::Spec
     end
 
     validation do
-      required(:songs).each do
-        schema do
-          required(:title).filled
+      params do
+        required(:songs).each do
+          schema do
+            required(:title).filled
+          end
         end
       end
     end
@@ -137,9 +138,11 @@ end
 
 class DryValidationExplicitSchemaTest < Minitest::Spec
   Session = Struct.new(:name, :email)
-  SessionSchema = Dry::Validation.Schema do
-    required(:name).filled
-    required(:email).filled
+  SessionSchema = Dry::Validation.Contract do
+    params do
+      required(:name).filled
+      required(:email).filled
+    end
   end
 
   class SessionForm < TestForm
@@ -175,28 +178,25 @@ class DryValidationDefaultGroupTest < Minitest::Spec
     property :email
     property :password
     property :confirm_password
-    property :starts_at, type: DRY_TYPES_CONSTANT::DateTime
-    property :active, type: DRY_TYPES_CONSTANT::Bool
+    property :starts_at, type: Types::Params::DateTime
+    property :active, type: Types::Params::Bool
     property :color
 
     validation do
-      required(:username).filled
-      required(:email).filled
-      required(:starts_at).filled(:date_time?)
-      required(:active).filled(:bool?)
+      params do
+        required(:username).filled
+        required(:email).filled
+        required(:starts_at).filled(:date_time?)
+        required(:active).filled(:bool?)
+      end
     end
 
     validation name: :another_block do
-      required(:confirm_password).filled
+      params { required(:confirm_password).filled }
     end
 
     validation name: :dynamic_args, with: {form: true} do
-      configure do
-        def colors
-          form.colors
-        end
-      end
-      required(:color).maybe(included_in?: colors)
+      params { required(:color).maybe(included_in?: form.colors) }
     end
 
     def colors
@@ -208,21 +208,27 @@ class DryValidationDefaultGroupTest < Minitest::Spec
 
   # valid.
   it do
-    form.validate(username: "Helloween",
-                  email:    "yep",
-                  starts_at: "01/01/2000 - 11:00",
-                  active: "true",
-                  confirm_password: "pA55w0rd").must_equal true
+    form.validate(
+      username: "Helloween",
+      email:    "yep",
+      starts_at: "01/01/2000 - 11:00",
+      active: "true",
+      confirm_password: "pA55w0rd"
+    ).must_equal true
+    form.active.must_equal true
     form.errors.messages.inspect.must_equal "{}"
   end
 
   it "invalid" do
-    form.validate(username: "Helloween",
-                  email:    "yep",
-                  active: "hello",
-                  starts_at: "01/01/2000 - 11:00",
-                  color: "purple").must_equal false
-    form.errors.messages.inspect.must_equal "{:active=>[\"must be boolean\"], :confirm_password=>[\"must be filled\"], :color=>[\"must be one of: red orange green\"]}"
+    form.validate(
+      username: "Helloween",
+      email:    "yep",
+      active: "1",
+      starts_at: "01/01/2000 - 11:00",
+      color: "purple"
+    ).must_equal false
+    form.active.must_equal true
+    form.errors.messages.inspect.must_equal "{:confirm_password=>[\"must be filled\"], :color=>[\"must be one of: red orange green\"]}"
   end
 end
 
@@ -239,21 +245,23 @@ class ValidationGroupsTest < MiniTest::Spec
       property :special_class
 
       validation do
-        required(:username).filled
-        required(:email).filled
-        required(:special_class).filled(type?: SomeClass)
+        params do
+          required(:username).filled
+          required(:email).filled
+          required(:special_class).filled(type?: SomeClass)
+        end
       end
 
       validation name: :email, if: :default do
-        required(:email).filled(min_size?: 3)
+        params { required(:email).filled(min_size?: 3) }
       end
 
       validation name: :password, if: :email do
-        required(:password).filled(min_size?: 2)
+        params { required(:password).filled(min_size?: 2) }
       end
 
       validation name: :confirm, if: :default, after: :email do
-        required(:confirm_password).filled(min_size?: 2)
+        params { required(:confirm_password).filled(min_size?: 2) }
       end
     end
 
@@ -301,12 +309,9 @@ class ValidationGroupsTest < MiniTest::Spec
         property :username
 
         validation name: :default, with: {user: OpenStruct.new(name: "Nick")} do
-          configure do
-            def users_name
-              user.name
-            end
+          params do
+            required(:username).filled(eql?: user.name)
           end
-          required(:username).filled(eql?: users_name)
         end
       end
 
@@ -331,14 +336,8 @@ class ValidationGroupsTest < MiniTest::Spec
   describe "with custom schema" do
     Session2 = Struct.new(:username, :email, :password)
 
-    MySchema = Dry::Validation.Schema do
-      configure do
-        config.messages_file = "test/fixtures/dry_error_messages.yml"
-
-        def good_musical_taste?(val)
-          val.is_a? String
-        end
-      end
+    MySchema = Dry::Schema.Params do
+      config.messages.load_paths << "test/fixtures/dry_error_messages.yml"
 
       required(:password).filled(min_size?: 6)
     end
@@ -349,8 +348,14 @@ class ValidationGroupsTest < MiniTest::Spec
       property :password
 
       validation schema: MySchema do
-        required(:username).filled
-        required(:email).filled(:good_musical_taste?)
+        params do
+          required(:username).filled
+          required(:email).filled
+        end
+
+        rule(:email) do
+          key.failure(:good_musical_taste?) unless value.is_a? String
+        end
       end
     end
 
@@ -358,17 +363,20 @@ class ValidationGroupsTest < MiniTest::Spec
 
     # valid.
     it do
+      skip "waiting dry-v to add this as feature https://github.com/dry-rb/dry-schema/issues/33"
       form.validate(username: "Helloween", email: "yep", password: "extrasafe").must_equal true
       form.errors.messages.inspect.must_equal "{}"
     end
 
     # invalid.
     it do
+      skip "waiting dry-v to add this as feature https://github.com/dry-rb/dry-schema/issues/33"
       form.validate({}).must_equal false
       form.errors.messages.must_equal(password: ["must be filled", "size cannot be less than 6"], username: ["must be filled"], email: ["must be filled", "you're a bad person"])
     end
 
     it do
+      skip "waiting dry-v to add this as feature https://github.com/dry-rb/dry-schema/issues/33"
       form.validate(email: 1).must_equal false
       form.errors.messages.inspect.must_equal "{:password=>[\"must be filled\", \"size cannot be less than 6\"], :username=>[\"must be filled\"], :email=>[\"you're a bad person\"]}"
     end
@@ -382,7 +390,7 @@ class ValidationGroupsTest < MiniTest::Spec
         property :title
 
         validation do
-          required(:title).filled
+          params { required(:title).filled }
         end
       end
 
@@ -390,7 +398,7 @@ class ValidationGroupsTest < MiniTest::Spec
         property :title
 
         validation do
-          required(:title).filled
+          params { required(:title).filled }
         end
       end
 
@@ -407,28 +415,23 @@ class ValidationGroupsTest < MiniTest::Spec
       end
 
       validation do
-        configure do
-          config.messages_file = "test/fixtures/dry_error_messages.yml"
-          # message need to be defined on fixtures/dry_error_messages
-          # d-v expects you to define your custome messages on the .yml file
-          def good_musical_taste?(value)
-            value != "Nickelback"
-          end
-        end
-
-        required(:title).filled(:good_musical_taste?)
-
-        required(:band).schema do
-          required(:name).filled
-          required(:label).schema do
-            required(:location).filled
-          end
-        end
-
-        required(:producers).each do
-          schema do
+        config.messages.load_paths << "test/fixtures/dry_new_api_error_messages.yml"
+        params do
+          required(:title).filled
+          required(:band).hash do
             required(:name).filled
+            required(:label).hash do
+              required(:location).filled
+            end
           end
+
+          required(:producers).each do
+            hash { required(:name).filled }
+          end
+        end
+
+        rule(:title) do
+          key.failure(:good_musical_taste?) unless value != "Nickelback"
         end
       end
     end
@@ -485,7 +488,7 @@ class ValidationGroupsTest < MiniTest::Spec
       form.band.errors.full_messages.must_equal ["Name must be filled", "Label Location must be filled"]
       form.band.label.errors.full_messages.must_equal ["Location must be filled"]
       form.producers.first.errors.full_messages.must_equal ["Name must be filled"]
-      form.errors.full_messages.must_equal ["Title must be filled", "Title you're a bad person", "Hit Title must be filled", "Songs Title must be filled", "Producers Name must be filled", "Band Name must be filled", "Band Label Location must be filled"]
+      form.errors.full_messages.must_equal ["Title must be filled", "Hit Title must be filled", "Songs Title must be filled", "Producers Name must be filled", "Band Name must be filled", "Band Label Location must be filled"]
     end
 
     describe "only 1 nested validation" do
@@ -499,16 +502,16 @@ class ValidationGroupsTest < MiniTest::Spec
         end
 
         validation do
-          configure do
-            config.messages_file = "test/fixtures/dry_error_messages.yml"
-          end
+          config.messages.load_paths << "test/fixtures/dry_new_api_error_messages.yml"
 
-          required(:title).filled
+          params do
+            required(:title).filled
 
-          required(:band).schema do
-            required(:name).filled
-            required(:label).schema do
-              required(:location).filled
+            required(:band).schema do
+              required(:name).filled
+              required(:label).schema do
+                required(:location).filled
+              end
             end
           end
         end
@@ -518,9 +521,9 @@ class ValidationGroupsTest < MiniTest::Spec
 
       it "allows to access dry's result semantics per nested form" do
         form.validate(
-          "title"  => "",
-          "songs"  => [{"title" => ""}, {"title" => ""}],
-          "band"   => {"size" => "", "label" => {"name" => ""}},
+          "title" => "",
+          "songs" => [{"title" => ""}, {"title" => ""}],
+          "band" => {"size" => "", "label" => {"name" => ""}},
           "producers" => [{"name" => ""}, {"name" => ""}, {"name" => "something lovely"}]
         )
 
@@ -573,17 +576,17 @@ class ValidationGroupsTest < MiniTest::Spec
       property :full_name, virtual: true
 
       validation name: :username do
-        configure do
-          config.messages_file = "test/fixtures/dry_error_messages.yml"
+        params do
+          required(:username).filled
+          required(:full_name).filled
         end
-
-        required(:username).filled
-        required(:full_name).filled
       end
 
-      validation name: :username, inherit: true do # overrides and extends the above.
-        optional(:username).maybe
-        required(:email).filled
+      validation name: :username, inherit: true do # extends the above.
+        params do
+          optional(:username).maybe(:string)
+          required(:email).filled
+        end
       end
     end
 
@@ -591,11 +594,13 @@ class ValidationGroupsTest < MiniTest::Spec
 
     # valid.
     it do
-      form.validate(full_name: "My name", email: 9).must_equal true
+      skip "waiting dry-v to add this as feature https://github.com/dry-rb/dry-schema/issues/33"
+      form.validate(email: 9).must_equal true
     end
 
     # invalid.
     it do
+      skip "waiting dry-v to add this as feature https://github.com/dry-rb/dry-schema/issues/33"
       form.validate({}).must_equal false
       form.errors.messages.must_equal email: ["must be filled"], full_name: ["must be filled"]
     end
@@ -608,17 +613,17 @@ class ValidationGroupsTest < MiniTest::Spec
       property :password
 
       validation name: :email do
-        required(:email).filled
+        params { required(:email).filled }
       end
 
       # run this is :email group is true.
       validation name: :after_email, if: ->(results) { results[:email].success? } do # extends the above.
-        required(:username).filled
+        params { required(:username).filled }
       end
 
       # block gets evaled in form instance context.
       validation name: :password, if: ->(results) { email == "john@trb.org" } do
-        required(:password).filled
+        params { required(:password).filled }
       end
     end
 
@@ -637,7 +642,7 @@ class ValidationGroupsTest < MiniTest::Spec
   end
 
   class NestedSchemaValidationTest < MiniTest::Spec
-    AddressSchema = Dry::Validation.Schema do
+    AddressSchema = Dry::Schema.Params do
       required(:company).filled(:int?)
     end
 
@@ -647,7 +652,7 @@ class ValidationGroupsTest < MiniTest::Spec
       end
 
       validation do
-        required(:delivery_address).schema(AddressSchema)
+        params { required(:delivery_address).schema(AddressSchema) }
       end
     end
 
@@ -666,7 +671,7 @@ class ValidationGroupsTest < MiniTest::Spec
       property :company
 
       validation do
-        required(:company).filled(:int?)
+        params { required(:company).filled(:int?) }
       end
     end
 
@@ -697,7 +702,7 @@ class ValidationGroupsTest < MiniTest::Spec
         property :enabled
 
         validation do
-          required(:title).filled
+          params { required(:title).filled }
         end
       end
 
@@ -707,20 +712,20 @@ class ValidationGroupsTest < MiniTest::Spec
       end
 
       validation do
-        configure do
-          config.messages_file = "test/fixtures/dry_error_messages.yml"
+        config.messages.load_paths << "test/fixtures/dry_new_api_error_messages.yml"
 
-          def a_song?(value)
-            value.any? { |el| el && el[:enabled] }
-          end
-
-          def with_last_name?(value)
-            !value[:last_name].nil?
-          end
+        params do
+          required(:songs).filled
+          required(:artist).filled
         end
 
-        required(:songs).filled(:a_song?)
-        required(:artist).filled(:with_last_name?)
+        rule(:songs) do
+          key.failure(:a_song?) unless value.any? { |el| el && el[:enabled] }
+        end
+
+        rule(:artist) do
+          key.failure(:with_last_name?) unless value[:last_name]
+        end
       end
     end
 
@@ -741,6 +746,55 @@ class ValidationGroupsTest < MiniTest::Spec
         :artist => ["must have last name"],
         :"songs.title" => ["must be filled"]
       )
+    end
+  end
+
+  class DryVWithSchemaAndParams < MiniTest::Spec
+    Foo = Struct.new(:age)
+
+    class ParamsForm < TestForm
+      property :age
+
+      validation do
+        params { required(:age).value(:integer) }
+
+        rule(:age) { key.failure("value exceeded") if value > 999 }
+      end
+    end
+
+    class SchemaForm < TestForm
+      property :age
+
+      validation do
+        schema { required(:age).value(:integer) }
+
+        rule(:age) { key.failure("value exceeded") if value > 999 }
+      end
+    end
+
+    it "using params" do
+      model = Foo.new
+      form = ParamsForm.new(model)
+      form.validate(age: "99").must_equal true
+      form.sync
+      model.age.must_equal "99"
+
+      form = ParamsForm.new(Foo.new)
+      form.validate(age: "1000").must_equal false
+      form.errors.messages.must_equal age: ["value exceeded"]
+    end
+
+    it "using schema" do
+      model = Foo.new
+      form = SchemaForm.new(model)
+      form.validate(age: "99").must_equal false
+      form.validate(age: 99).must_equal true
+      form.sync
+      model.age.must_equal 99
+
+      form = SchemaForm.new(Foo.new)
+      form.validate(age: 1000).must_equal false
+      form.errors.messages.must_equal age: ["value exceeded"]
     end
   end
